#76 refactor: move some service methods to core library
